+++ OctreeApi/OctreeLibrary/src/sparseOctree/debugging.cu
@@ -77,15 +77,15 @@ template unordered_map<uint32_t, GpuArrayU32> const& SparseOctree<float, uint8_t
 template uint32_t SparseOctree<float, uint8_t>::getRootIndex ();
 
 //----------------------------------------------------------------------------------------------------------------------
-//                                           SparseOctree<double, uint16_t>
+//                                           SparseOctree<double, uint8_t>
 //----------------------------------------------------------------------------------------------------------------------
 
-template const OctreeMetadata& SparseOctree<double, uint16_t>::getMetadata () const;
-template unique_ptr<uint32_t[]> SparseOctree<double, uint16_t>::getDataLUT () const;
-template unique_ptr<uint32_t[]> SparseOctree<double, uint16_t>::getDensePointCountPerVoxel () const;
-template unique_ptr<int[]> SparseOctree<double, uint16_t>::getDenseToSparseLUT () const;
-template unique_ptr<int[]> SparseOctree<double, uint16_t>::getSparseToDenseLUT () const;
-template unique_ptr<Chunk[]> SparseOctree<double, uint16_t>::getOctreeSparse () const;
-template unordered_map<uint32_t, GpuArrayU32> const& SparseOctree<double, uint16_t>::
+template const OctreeMetadata& SparseOctree<double, uint8_t>::getMetadata () const;
+template unique_ptr<uint32_t[]> SparseOctree<double, uint8_t>::getDataLUT () const;
+template unique_ptr<uint32_t[]> SparseOctree<double, uint8_t>::getDensePointCountPerVoxel () const;
+template unique_ptr<int[]> SparseOctree<double, uint8_t>::getDenseToSparseLUT () const;
+template unique_ptr<int[]> SparseOctree<double, uint8_t>::getSparseToDenseLUT () const;
+template unique_ptr<Chunk[]> SparseOctree<double, uint8_t>::getOctreeSparse () const;
+template unordered_map<uint32_t, GpuArrayU32> const& SparseOctree<double, uint8_t>::
         getSubsampleLUT () const;
-template uint32_t SparseOctree<double, uint16_t>::getRootIndex ();
+template uint32_t SparseOctree<double, uint8_t>::getRootIndex ();
